requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
openai-whisper>=20231117
imageio-ffmpeg==0.6.0
pdf2docx>=0.5.8
python-docx>=0.8.11
paddleocr==2.7.3
numpy==1.26.4
pyyaml>=6.0
aiofiles>=23.0.0
